+++ ui/packages/shared/components/src/index.tsx
@@ -28,9 +28,9 @@ import ParcaContext from './ParcaContext';
 import Pill, {PillVariant} from './Pill';
 import ResponsiveSvg from './ResponsiveSvg';
 import Select, {type SelectElement, type SelectItem} from './Select';
-import FlamegraphSkeleton from './Skeletons/FlamegraphSkeleton';
-import IcicleGraphSkeleton, {IcicleActionButtonPlaceholder} from './Skeletons/IcicleGraphSkeleton';
+import FlameGraphSkeleton, {FlameActionButtonPlaceholder} from './Skeletons/FlamegraphSkeleton';
 import MetricsGraphSkeleton from './Skeletons/MetricsGraphSkeleton';
+import SandwichFlameGraphSkeleton from './Skeletons/SandwichFlameGraphSkeleton';
 import SourceSkeleton from './Skeletons/SourceSkeleton';
 import TableSkeleton, {TableActionButtonPlaceholder} from './Skeletons/TableSkeleton';
 import Spinner from './Spinner';
@@ -55,11 +55,11 @@ export {
   DateTimeRange,
   DateTimeRangePicker,
   Dropdown,
-  FlamegraphSkeleton,
+  FlameGraphSkeleton,
   GrpcMetadataContext,
   GrpcMetadataProvider,
-  IcicleActionButtonPlaceholder,
-  IcicleGraphSkeleton,
+  FlameActionButtonPlaceholder,
+  SandwichFlameGraphSkeleton,
   IconButton,
   Input,
   KeyDownProvider,