+++ ui/packages/shared/profile/src/ProfileFlameGraph/FlameGraphArrow/ContextMenuWrapper.tsx
@@ -30,7 +30,7 @@ interface ContextMenuWrapperProps {
   hideMenu: () => void;
   hideBinary: (binaryToRemove: string) => void;
   unit?: string;
-  isSandwich?: boolean;
+  isInSandwichView?: boolean;
 }
 
 export interface ContextMenuWrapperRef {