+++ ui/packages/shared/profile/src/ProfileFlameGraph/FlameGraphArrow/FlameGraphNodes.tsx
@@ -20,8 +20,6 @@ import {selectBinaries, useAppSelector} from '@parca/store';
 
 import 'react-contexify/dist/ReactContexify.css';
 
-import {USER_PREFERENCES} from '@parca/hooks';
-
 import {ProfileSource} from '../../ProfileSource';
 import TextWithEllipsis from './TextWithEllipsis';
 import {
@@ -31,7 +29,6 @@ import {
   FIELD_FUNCTION_FILE_NAME,
   FIELD_FUNCTION_NAME,
   FIELD_MAPPING_FILE,
-  FIELD_PARENT,
   FIELD_TIMESTAMP,
   FIELD_VALUE_OFFSET,
 } from './index';
@@ -44,7 +41,7 @@ export interface colorByColors {
   [key: string]: string;
 }
 
-export interface IcicleNodeProps {
+export interface FlameNodeProps {
   height: number;
   totalWidth: number;
   table: Table<any>;
@@ -57,30 +54,30 @@ export interface IcicleNodeProps {
   colorForSimilarNodes: string;
   selectedRow: number;
   onClick: () => void;
-  isIcicleChart: boolean;
+  isFlameChart: boolean;
   profileSource: ProfileSource;
-  isFlamegraph?: boolean;
-  isSandwich?: boolean;
+  isRenderedAsFlamegraph?: boolean;
+  isInSandwichView?: boolean;
   maxDepth?: number;
   effectiveDepth?: number;
   tooltipId?: string;
 
-  // Hovering row must only ever be used for highlighting similar nodes, otherwise it will cause performance issues as it causes the full iciclegraph to get rerendered every time the hovering row changes.
+  // Hovering row must only ever be used for highlighting similar nodes, otherwise it will cause performance issues as it causes the full flamegraph to get rerendered every time the hovering row changes.
   hoveringRow?: number;
   setHoveringRow: (row: number | undefined) => void;
 }
 
-export const icicleRectStyles = {
+export const flameRectStyles = {
   cursor: 'pointer',
   transition: 'opacity .15s linear',
 };
-export const fadedIcicleRectStyles = {
+export const fadedFlameRectStyles = {
   cursor: 'pointer',
   transition: 'opacity .15s linear',
   opacity: '0.5',
 };
 
-export const IcicleNode = React.memo(function IcicleNodeNoMemo({
+export const FlameNode = React.memo(function FlameNodeNoMemo({
   table,
   row,
   colors,
@@ -95,14 +92,14 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
   onContextMenu,
   hoveringRow,
   setHoveringRow,
-  isIcicleChart,
+  isFlameChart,
   profileSource,
-  isFlamegraph = false,
-  isSandwich = false,
+  isRenderedAsFlamegraph = false,
+  isInSandwichView = false,
   maxDepth = 0,
   effectiveDepth,
   tooltipId = 'default',
-}: IcicleNodeProps): React.JSX.Element {
+}: FlameNodeProps): React.JSX.Element {
   // get the columns to read from
   const mappingColumn = table.getChild(FIELD_MAPPING_FILE);
   const functionNameColumn = table.getChild(FIELD_FUNCTION_NAME);
@@ -171,8 +168,8 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
     return <></>;
   }
 
-  if (row === 0 && (isIcicleChart || isSandwich)) {
-    // The root node is not rendered in the icicle chart or sandwich view, so we return null.
+  if (row === 0 && (isFlameChart || isInSandwichView)) {
+    // The root node is not rendered in the flame chart or sandwich view, so we return null.
     return <></>;
   }
 
@@ -180,7 +177,7 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
   const tsBounds = boundsFromProfileSource(profileSource);
   const total = cumulativeColumn?.get(selectedRow);
   const totalRatio = cumulative > total ? 1 : Number(cumulative) / Number(total);
-  const width: number = isIcicleChart
+  const width: number = isFlameChart
     ? (Number(cumulative) / (Number(tsBounds[1]) - Number(tsBounds[0]))) * totalWidth
     : totalRatio * totalWidth;
 
@@ -190,12 +187,12 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
 
   const selectedDepth = depthColumn?.get(selectedRow);
   const styles =
-    selectedDepth !== undefined && selectedDepth > depth ? fadedIcicleRectStyles : icicleRectStyles;
+    selectedDepth !== undefined && selectedDepth > depth ? fadedFlameRectStyles : flameRectStyles;
 
   const onMouseEnter = (): void => {
     setHoveringRow(row);
     window.dispatchEvent(
-      new CustomEvent(`icicle-tooltip-update-${tooltipId}`, {
+      new CustomEvent(`flame-tooltip-update-${tooltipId}`, {
         detail: {row},
       })
     );
@@ -204,7 +201,7 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
   const onMouseLeave = (): void => {
     setHoveringRow(undefined);
     window.dispatchEvent(
-      new CustomEvent(`icicle-tooltip-update-${tooltipId}`, {
+      new CustomEvent(`flame-tooltip-update-${tooltipId}`, {
         detail: {row: null},
       })
     );
@@ -216,25 +213,25 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
 
   const ts = tsColumn !== null ? Number(tsColumn.get(row)) : 0;
   const x =
-    isIcicleChart && tsColumn !== null
+    isFlameChart && tsColumn !== null
       ? ((ts - Number(tsBounds[0])) / (Number(tsBounds[1]) - Number(tsBounds[0]))) * totalWidth
       : selectedDepth > depth
       ? 0
       : ((Number(valueOffset) - Number(selectionOffset)) / Number(total)) * totalWidth;
 
   const calculateY = (
-    isFlamegraph: boolean,
-    isSandwich: boolean,
-    isIcicleChart: boolean,
+    isRenderedAsFlamegraph: boolean,
+    isInSandwichView: boolean,
+    isFlameChart: boolean,
     maxDepth: number,
     depth: number,
     height: number
   ): number => {
-    if (isFlamegraph) {
+    if (isRenderedAsFlamegraph) {
       return (maxDepth - depth) * height; // Flamegraph is inverted
     }
 
-    if (isIcicleChart || isSandwich) {
+    if (isFlameChart || isInSandwichView) {
       return (depth - 1) * height;
     }
 
@@ -242,9 +239,9 @@ export const IcicleNode = React.memo(function IcicleNodeNoMemo({
   };
 
   const y = calculateY(
-    isFlamegraph,
-    isSandwich,
-    isIcicleChart,
+    isRenderedAsFlamegraph,
+    isInSandwichView,
+    isFlameChart,
     effectiveDepth ?? maxDepth,
     depth,
     height