+++ ui/packages/shared/profile/src/ProfileSelector/useAutoQuerySelector.ts
@@ -99,7 +99,7 @@ export const useAutoQuerySelector = ({
       void navigateTo('/', {
         ...compareQuery,
         search_string: '',
-        dashboard_items: ['icicle'],
+        dashboard_items: ['flamegraph'],
       });
     }
   }, [comparing, querySelection, navigateTo, expressionA, dispatch, loading]);