+++ ui/packages/shared/client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.17.2 (2025-07-10)
+
+**Note:** Version bump only for package @parca/client
+
 ## 0.17.1 (2025-07-08)
 
 **Note:** Version bump only for package @parca/client