+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.102 (2025-07-10)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.101 (2025-07-08)
 
 **Note:** Version bump only for package @parca/utilities